0.0.2

+++ scan.rb
@@ -32,7 +32,7 @@ class Scanner
   def tika(stream, mimetype, size)
     req = Net::HTTP::Put.new('/all')
     req.body_stream = stream
-    #req["Content-Type"] = mimetype
+    req["Content-Type"] = mimetype
     req.add_field('Content-Length', size)
     req.add_field('Accept', 'application/x-tar')
     res = TIKA.request(req)
@@ -58,16 +58,11 @@ class Scanner
     return data
   end
 
-  def html2text(html)
-    return Nokogiri::HTML(html).text
-  end
-
   def text2words(text)
     text = text.strip.split(/[^[:alpha:]]+/)
     text = text.collect{|w| w.downcase}.uniq.reject{|w| w.length < 2}
     chars = 0
-    text = Hash[*(text.collect{|w| l = w.length; chars += l; [w, l]}.flatten)]
-    return [text, chars]
+    return Hash[*(text.collect{|w| l = w.length; chars += l; [w, l]}.flatten)]
   end
 
   def initialize(root)
@@ -79,69 +74,63 @@ class Scanner
     extensions = []
     hashes = {}
     errors = []
-    Dir[File.join(root, '*.zip')].sort.reverse.each{|zipfile|
+    Dir[File.join(root, '*.zip')].sort.each{|zipfile|
       key = File.basename(zipfile, File.extname(zipfile)).upcase
-      puts key
+      cachefile = File.join(File.dirname(zipfile), ".#{key}#{EXT}")
+      next if File.exists?(cachefile) && File.mtime(cachefile) > File.mtime(zipfile)
+
 
       propfile = File.join(File.dirname(zipfile), File.basename(zipfile, File.extname(zipfile)) + '.prop')
-
       hash = nil
       File.open(propfile){|f| hash = Nokogiri::XML(f).xpath('/properties/hash').inner_text }
       hash = nil if hash == ''
-
       hashes[key] = hash if hash
-
-      cachefile = File.join(File.dirname(zipfile), ".#{key}#{EXT}")
-      update = !File.exists?(cachefile) || File.mtime(zipfile) > File.mtime(cachefile)
 
       data = {text: ''}
 
-      if update
-        File.unlink(cachefile) if File.exists?(cachefile)
+      File.unlink(cachefile) if File.exists?(cachefile)
 
-        Zip::ZipFile.foreach(zipfile) {|entry|
-          next unless entry.file?
+      puts zipfile
+      Zip::ZipFile.foreach(zipfile) {|entry|
+        next unless entry.file?
 
-          ext = File.extname(entry.to_s.downcase)
+        ext = File.extname(entry.to_s.downcase)
 
-          case ext
-          when '.txt'
-            data = {text: entry.get_input_stream.read}
-          when '.html'
-            data[:html] ||= ''
-            data[:html] += ' ' + entry.get_input_stream.read
-          when '.epub'
-            # Tika doesn't always pick up epub contents
-            data[:html] ||= ''
-            Tempfile.open('epub') do |f|
-              entry.extract(f.path){true}
-              Zip::ZipFile.foreach(f.path){|chapter|
-                next unless chapter.file?
-                next unless chapter.to_s.downcase =~ /\.html?$/ || (chapter.to_s.downcase =~ /oebps\/text/ && chapter.to_s.downcase =~ /\.xml$/)
-                data[:html] += ' ' + chapter.get_input_stream.read
-              }
-            end
-          when '.pdf', '.mobi', '.doc', '.docx'
-            raise "No mimetype for #{ext}" unless EXT2MIMETYPE[ext]
-            data = tika(entry.get_input_stream, EXT2MIMETYPE[ext], entry.size)
-          else
-            extensions << File.extname(entry.to_s.downcase)
+        case ext
+        when '.txt'
+          data = {text: entry.get_input_stream.read}
+        when '.html'
+          data[:text] += ' ' + Nokogiri::HTML(entry.get_input_stream.read).inner_text
+        when '.epub'
+          # Tika doesn't always pick up epub contents
+          Tempfile.open('epub') do |f|
+            entry.extract(f.path){true}
+            Zip::ZipFile.foreach(f.path){|chapter|
+              next unless chapter.file?
+              next unless chapter.to_s.downcase =~ /\.x?html?$/ || (chapter.to_s.downcase =~ /oebps\/text/ && chapter.to_s.downcase =~ /\.xml$/)
+              data[:text] += ' ' +  Nokogiri::HTML(chapter.get_input_stream.read).inner_text
+            }
           end
-        }
-
-        data[:text] = html2text(data.delete(:html)) if data[:html]
-        if data[:text].to_s.strip == ''
-          errors << zipfile
-          next
+        when '.pdf', '.mobi', '.doc', '.docx'
+          raise "No mimetype for #{ext}" unless EXT2MIMETYPE[ext]
+          data = tika(entry.get_input_stream, EXT2MIMETYPE[ext], entry.size)
+          data[:text] = Nokogiri::HTML(data.delete(:html) || '').inner_text
+        else
+          extensions << File.extname(entry.to_s.downcase)
         end
+      }
 
-        data[:chars] = data[:text].length
-        data[:words] = text2words(data.delete(:text))
-        data[:pages] = data[:metadata]['xmpTPg:NPages'] if data[:metadata] && data[:metadata]['xmpTPg:NPages']
-        data.delete(:metadata)
-
-        File.open(cachefile, "wb", :encoding => 'utf-8'){|f| f.write(data.to_json) } if data[:words].size > 0
+      if data[:text].to_s.strip == ''
+        errors << zipfile
+        next
       end
+
+      data[:chars] = data[:text].length
+      data[:words] = text2words(data.delete(:text))
+      data[:pages] = data[:metadata]['xmpTPg:NPages'] if data[:metadata] && data[:metadata]['xmpTPg:NPages']
+      data.delete(:metadata)
+
+      File.open(cachefile, "wb", :encoding => 'utf-8'){|f| f.write(data.to_json) } if data[:words].size > 0
     }
 
     File.open(File.join(root, EXT), "wb", :encoding => 'utf-8'){|f| f.write(hashes.to_json) }
